Revert "fixed possible overflow InventoryMenu"
This reverts commit bac937aae40d04d774ac48d33ac67e7224c6a064.

+++ Game/world/item.cpp
@@ -172,7 +172,7 @@ int32_t Item::uiValue(size_t id) const {
   return hitem.count[id];
   }
 
-uint32_t Item::count() const {
+size_t Item::count() const {
   return hitem.amount;
   }
